requirements.txt
nltk
torch
safetensors
tqdm
sentence-transformers
python-Levenshtein
pandas
numpy
retsim-pytorch